ColorTypes kan nu ses, slettes og laves og ændres (næsten)

+++ src/pages/color/colorTypes/colorTypeFunctions.js
@@ -2,14 +2,36 @@ var URL = "http://localhost:8080/api/color-types"
 let router;
 
 export function initColorTypes(navigoRouter) {
-
-    document.getElementById("submit").onclick = addColorType;
     getAllColorTypes();
     router = navigoRouter
+    document.getElementById("submit").onclick = addColorType;
     document.getElementById("table").onclick = (element) =>{
         let id = element.target.id
-        deleteColorType(id);
+        if(id.includes("delete")) {
+            deleteColorType(id);
+        }
     }
+    document.getElementById("table").ondblclick = (element) =>{
+        let id = element.target.id
+        if(id.includes("text")){
+            changeTdToInput(id);
+        }
+        var el = document.getElementById(id);
+        el.addEventListener("keydown", function(event) {
+            if (event.key === "Enter") {
+                console.log("Enter clicked "  + id)
+                let field = document.getElementById(id)
+                field.readOnly = true
+                editColorType(id)
+            }
+        });
+    }
+
+}
+
+function changeTdToInput(id){
+    const field = document.getElementById(id)
+    field.readOnly = false
 }
 
 
@@ -18,13 +40,12 @@ async function getAllColorTypes() {
     document.getElementById("tbody-all").innerHTML = ""
     try{
         const data = await fetch(URL).then(res => res.json());
-        console.log(data)
         const tableRowsArray = data.map(
             (colorType) =>
                 `
         <tr>
-            <td>${colorType.id}</td>
-            <td>${colorType.type}</td>
+            <td id="color-type-id-${colorType.id}" value="${colorType.id}">${colorType.id}</td>
+            <td><input readonly type='text' id="text${colorType.id}" value='${colorType.type}'></td>
             <td><button id="delete${colorType.id}">Delete</button></td>
         `
         );
@@ -38,7 +59,6 @@ async function getAllColorTypes() {
 
 async function addColorType() {
     const type = document.getElementById("if1").value;
-    console.log(type)
     const newColorType = {
         type
     };
@@ -55,23 +75,38 @@ async function addColorType() {
 
 }
 
-
-
-
 async function deleteColorType(idToDelete) {
-    console.log(idToDelete)
-    if(idToDelete.includes("delete")){
-        idToDelete = idToDelete.split('delete')[1]
-        const response = await fetch(URL + "/" + idToDelete, {
-            method: "DELETE",
+    idToDelete = idToDelete.split('delete')[1]
+    const response = await fetch(URL + "/" + idToDelete, {
+        method: "DELETE",
 
-        }).then((res) => res.json())
-        console.log(response)
-        location.reload();
-    }
+    }).then((res) => res.json())
+    location.reload();
+}
 
+async function editColorType(id) {
+    const idToDelete = id.split('text')[1]
+    const colorTypeId = document.getElementById("color-type-id-"+idToDelete).innerText
+    const colorTypeText = document.getElementById(id).value;
+
+    const editedColorType = {
+        colorTypeId,
+        colorTypeText
+    };
+    console.log(colorTypeId + " id")
+    console.log(editedColorType)
+    const data = await fetch(URL, {
+        method: "PUT",
+        headers: {
+            "Content-Type": "application/json",
+        },
+        body: JSON.stringify(editedColorType),
+    })
+        .then((res) => res.json())
 
 }
+
+
